// client/src/app/_models/userParams.ts
import { User } from "./user";

export class UserParams {
    gender: string;
    minAge = 10;
    maxAge = 99
    pageNumber = 1;
    pageSize = 5;
    orderBy = 'lastActive';

   
    constructor(user:User | null) {
       this.gender = user?.gender === 'female' ? 'male' :'female'
        
    }

}   